requirements.txt
aniso8601==8.0.0
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.20.0
pycryptodome==3.9.8
pymongo==3.11.0
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
